Implement solution for LeetCode problem 169
-Majority Element

This commit introduces Java solutions for the "Majority Element" problem on LeetCode. The algorithm addresses the task of finding the element that appears more than ⌊n/2⌋ times in a given array, where n is the size of the array. Three different approaches are implemented and tested for efficiency and correctness.